ToDo: get the items synchronously to avoid a flicker
