+++ extensions/todo-list/src/index.tsx
@@ -1,9 +1,7 @@
-import { ActionPanel, clearSearchBar, environment, List, showToast, ToastStyle } from "@raycast/api";
-import { useEffect, useState } from "react";
-import fs from "fs/promises";
+import { ActionPanel, clearSearchBar, List, showToast, ToastStyle } from "@raycast/api";
+import { useState } from "react";
 import { searchModeAtom, todoAtom } from "./atoms";
 import { useAtom } from "jotai";
-import { DEFAULT_SECTIONS, TODO_FILE } from "./config";
 import _ from "lodash";
 import { insertIntoSection, compare } from "./utils";
 import DeleteAllAction from "./delete_all";
@@ -13,42 +11,7 @@ import SearchModeAction from "./search_mode_action";
 export default function TodoList() {
   const [todoSections, setTodoSections] = useAtom(todoAtom);
   const [newTodoText, setNewTodoText] = useState("");
-  const [searchMode, setSearchMode] = useAtom(searchModeAtom);
-  const [loading, setLoading] = useState(true);
-  useEffect(() => {
-    (async () => {
-      try {
-        const storedItemsBuffer = await fs.readFile(TODO_FILE);
-        const storedItems = JSON.parse(storedItemsBuffer.toString());
-        // from v1 where items were stored in an array
-        if (Array.isArray(storedItems)) {
-          const storedPinned = storedItems[0];
-          const storedTodo = [];
-          const storedCompleted = [];
-          for (const todo of storedItems[1]) {
-            if (todo.completed) {
-              storedCompleted.push(todo);
-            } else {
-              storedTodo.push(todo);
-            }
-          }
-          const convertedStoredItems = {
-            pinned: storedPinned,
-            todo: storedTodo,
-            completed: storedCompleted,
-          };
-          setTodoSections(convertedStoredItems);
-        } else {
-          setTodoSections(storedItems);
-        }
-      } catch (error) {
-        await fs.mkdir(environment.supportPath, { recursive: true });
-        setTodoSections(DEFAULT_SECTIONS);
-      } finally {
-        setLoading(false);
-      }
-    })();
-  }, []);
+  const [searchMode] = useAtom(searchModeAtom);
 
   const addTodo = async () => {
     if (newTodoText.length === 0) {
@@ -72,7 +35,6 @@ export default function TodoList() {
   return (
     <List
       key={searchMode ? "search" : "nosearch"}
-      isLoading={loading}
       actions={
         <ActionPanel>
           {!searchMode && <ActionPanel.Item title="Create Todo" onAction={() => addTodo()} />}
